Migrated S3 code to new class
