+++ runner/src/streamer-message-fetcher/index.ts
@@ -1,0 +1,1 @@
+export { default } from './s3-streamer-fetcher';
